Fazendo ajustes finos para padronização

+++ lib/models/products.dart
@@ -2,7 +2,7 @@ class Product {
   int id;
   String nomeProduto;
   int quantidade;
-  int preco;
+  double preco;
 
   Product({
   required this.id,
